% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{add_adapters}
\alias{add_adapters}
\title{Add an adapter functions to an \link[=new_adapters]{adapters} or a
\link[=new_file_structure_fwf]{pecification}
or a \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection} class object}
\usage{
add_adapters(obj, adapters)
}
\arguments{
\item{obj}{This argument can be various different things:
\itemize{
\item An \link[=new_adapters]{adapters} class object. Then the \code{adapters} will be
appended to this object.
\item An \link[=new_file_structure_fwf]{file_structure} class object. Then the \code{adapters}
will be appended to the \code{obj$adapters} attribute of the class.
\item An \link[=new_file_collection]{file_collection} class object. Then the \code{adapters}
will be appended to each \link[=new_file_definition]{file_definition} object in the
list.
}}

\item{adapters}{An \link[=new_adapters]{adapters} class object, which holds
all adapter functions and their descriptions. This class object can be
created by using the function \code{\link[=new_adapters]{new_adapters()}}.}
}
\value{
The modified object
}
\description{
Add an adapter functions to an \link[=new_adapters]{adapters} or a
\link[=new_file_structure_fwf]{pecification}
or a \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection} class object
}
\section{adapters}{

An adapter function is a function that takes a data.frame as input argument
and returns a modified version of this data.frame.
The adapter functions are stored in an \link[=new_adapters]{adapters}
class object, which is a special list that contains all adapter functions
and a description text of each function. This class objects can be
created by using the function \code{\link[=new_adapters]{new_adapters()}}.
The \link[=new_adapters]{adapters} class objects can be added to a
\link[=new_file_structure_fwf]{file_structure} or a
\link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection} class object.
After reading a data file (by calling \link[=read_data]{read_data(file_definition)})
all adapter functions listed in the \code{adapters} argument of the
file_definition]\code{\link[=new_file_definition]{new_file_definition()}} class object
will be applied consecutively to the loaded data set.
Adapter functions can be added to an existing
\link[=new_file_structure_fwf]{file_structure} or a \link[=new_file_definition]{file_definition} or
a \link[=new_file_collection]{file_collection} class
object by using the function \code{\link[=add_adapters]{add_adapters()}}.
Adapter functions can be used for several tasks:
\itemize{
\item adapt the data sets in such a way that they can be concatenated for
mutliple years
\item compute new variables from existing variables
\item fix errors in variables
\item transform the values of a variable of an older data set, such that it
complies with a newer variable definition
}
}

