% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{add_adapters_.default}
\alias{add_adapters_.default}
\title{Missing class for \code{\link[=add_adapters_]{add_adapters_()}}}
\usage{
\method{add_adapters_}{default}(
  obj,
  adapters = NULL,
  err_h = composerr("Error while calling 'add_adapters_()'")
)
}
\arguments{
\item{obj}{This argument can be various different things:
\itemize{
\item An \link[=new_adapters]{adapters} class object. Then the \code{adapters} will be
appended to this object.
\item An \link[=new_file_structure_fwf]{file_structure} class object. Then the \code{adapters}
will be appended to the \code{obj$adapters} attribute of the class.
\item An \link[=new_file_collection]{file_collection} class object. Then the \code{adapters}
will be appended to each \link[=new_file_definition]{file_definition} object in the
list.
}}

\item{adapters}{An \link[=new_adapters]{adapters} class object, which holds
all adapter functions and their descriptions. This class object can be
created by using the function \code{\link[=new_adapters]{new_adapters()}}.}

\item{err_h}{An error handler}
}
\description{
Missing class for \code{\link[=add_adapters_]{add_adapters_()}}
}
