% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_character_vector}
\alias{check_character_vector}
\title{Check that a specific list element is a character vector}
\usage{
check_character_vector(
  obj,
  attr,
  err_h,
  allowed_values = NULL,
  allow_na = FALSE,
  allow_null = FALSE,
  allow_na_last = FALSE
)
}
\arguments{
\item{obj}{The list object, that should be checked}

\item{attr}{A string holding the name of the list element, that should be
checked.}

\item{err_h}{An error handler.}

\item{allowed_values}{An optional set of allowed strings.}

\item{allow_na}{Are \code{NA} values allowed?}

\item{allow_null}{A logical, defining whether the argument is allowed to be \code{NULL}.}

\item{allow_na_last}{A logical value, defining if the last entry of the
vector is allowed to be \code{NA}.}
}
\description{
Assumes non-zero length.
}
