% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_excel_range}
\alias{check_excel_range}
\title{Check that a specific list element is an excel range string}
\usage{
check_excel_range(obj, attr, err_h, allow_na = FALSE, allow_null = FALSE)
}
\arguments{
\item{obj}{The list object, that should be checked}

\item{attr}{A string holding the name of the list element, that should be
checked.}

\item{err_h}{An error handler.}

\item{allow_na}{Are \code{NA} values allowed?}

\item{allow_null}{A logical, defining whether the argument is allowed to be \code{NULL}.}
}
\description{
Assumes non-zero length.
}
