% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_list}
\alias{check_list}
\title{Check that object is a list object}
\usage{
check_list(
  obj,
  err_h,
  allow_null = FALSE,
  allow_names = TRUE,
  allow_empty = TRUE
)
}
\arguments{
\item{obj}{The object, that should be checked.}

\item{err_h}{An error handler.}

\item{allow_null}{A logical, defining whether the object is allowed to be \code{NULL}.}

\item{allow_names}{A logical, defining whether the entries may have names.}

\item{allow_empty}{A logical, defining whether the list may be empty.}
}
\value{
The passed in object.
}
\description{
Check that object is a list object
}
