% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_missing}
\alias{check_missing}
\title{Check that a specific list element is missing}
\usage{
check_missing(obj, attr, err_h)
}
\arguments{
\item{obj}{The list object, that should be checked}

\item{attr}{A string holding the name of the list element, that should be
checked.}

\item{err_h}{An error handler.}
}
\description{
Attribute is expected to be missing.
}
