% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_numeric}
\alias{check_numeric}
\title{Check that a specific list element is a numeric value}
\usage{
check_numeric(
  obj,
  attr,
  err_h,
  min_val = NULL,
  max_val = Inf,
  allow_na = FALSE,
  allow_null = FALSE,
  allow_inf = FALSE
)
}
\arguments{
\item{obj}{The list object, that should be checked}

\item{attr}{A string holding the name of the list element, that should be
checked.}

\item{err_h}{An error handler.}

\item{min_val}{An optional argument, defining a minimal allowed value.}

\item{max_val}{An numeric argument, defining the maximal allowed value.}

\item{allow_na}{Are \code{NA} values allowed?}

\item{allow_null}{A logical, defining whether the argument is allowed to be \code{NULL}.}

\item{allow_inf}{An optional logical, defining if the value \code{Inf} is allowed.}
}
\description{
Check that a specific list element is a numeric value
}
