% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_required_attributes}
\alias{check_required_attributes}
\title{Check that a list object has all required attributes}
\usage{
check_required_attributes(obj, attrs_required, err_h)
}
\arguments{
\item{obj}{The list object, that should be checked}

\item{attrs_required}{A character vector, holding the names of the required list elements}

\item{err_h}{An error handler.}
}
\description{
Check that a list object has all required attributes
}
