% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols.R
\name{cols_get_attribute}
\alias{cols_get_attribute}
\title{Extract column definitions from \code{cols} argument}
\usage{
cols_get_attribute(cols, attr)
}
\arguments{
\item{cols}{A list object, where each entry represents a column definition.
Each element of the list, is a list itself with possible entries:
\itemize{
\item \code{name}: Column name (optional)
\item \code{type}: Colunn type (required)
\item \code{start}: Column start postition (optional and only for FWF)
\item \code{end}: Column end postition (optional and only for FWF)
\item \code{width}: Column width (optional and only for FWF)
}}

\item{attr}{A character string, holding the name of the Definition-Attribute,
which should be pulled out.}
}
\value{
A character or integer vector
}
\description{
Extract column definitions from \code{cols} argument
}
