% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{create_meta_df}
\alias{create_meta_df}
\title{Helper function for creating the meta data data.frame in \code{\link[=get_meta]{get_meta()}}}
\usage{
create_meta_df(cols, col_names, col_types, file_meta)
}
\arguments{
\item{cols}{A numeric vector holding the column indices}

\item{col_names}{A character vector holding the column names}

\item{col_types}{A character vector holding the column types}

\item{file_meta}{A \link[=new_file_meta]{file_meta} class object holding the
meta information for all columns.}
}
\value{
A data.frame holding the meta information for columns specified in
\code{cols}.
}
\description{
Helper function for creating the meta data data.frame in \code{\link[=get_meta]{get_meta()}}
}
