% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\docType{data}
\name{enum_col_types}
\alias{enum_col_types}
\title{Possible column types}
\format{
An object of class \code{list} of length 5.
}
\usage{
enum_col_types
}
\description{
Possible column types
}
\keyword{datasets}
