% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\docType{data}
\name{enum_encodings}
\alias{enum_encodings}
\title{Possible file types}
\format{
An object of class \code{list} of length 2.
}
\usage{
enum_encodings
}
\description{
Possible file types
}
\keyword{datasets}
