% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\docType{data}
\name{enum_file_types}
\alias{enum_file_types}
\title{Possible file types}
\format{
An object of class \code{list} of length 4.
}
\usage{
enum_file_types
}
\description{
Possible file types
}
\keyword{datasets}
