% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file_definition.R
\name{err_file_definition}
\alias{err_file_definition}
\title{Throw an error: Object is not of class  \code{file_structure}}
\usage{
err_file_definition(err_h)
}
\arguments{
\item{err_h}{An error handler}
}
\description{
Throw an error: Object is not of class  \code{file_structure}
}
