% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_col_names.default}
\alias{get_col_names.default}
\title{Missing class for \code{\link[=get_col_names]{get_col_names()}}}
\usage{
\method{get_col_names}{default}(
  obj,
  err_h = composerr("Error while calling 'get_col_names()'"),
  ...
)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{err_h}{An error handler}

\item{...}{Optional error handler}
}
\description{
Missing class for \code{\link[=get_col_names]{get_col_names()}}
}
