% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_col_names}
\alias{get_col_names}
\alias{get_col_names.file_structure_fwf}
\alias{get_col_names.file_structure_dsv}
\alias{get_col_names.file_structure_excel}
\title{Helper function for \code{read_data()}: Generate \code{col_names} object}
\usage{
get_col_names(obj, ...)

\method{get_col_names}{file_structure_fwf}(obj, ...)

\method{get_col_names}{file_structure_dsv}(obj, ...)

\method{get_col_names}{file_structure_excel}(obj, ...)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{...}{Optional error handler}
}
\value{
A character vector holding the column names. If no column names
are defined, then the column names are replaced by the values \code{x1}, \code{x2}, ...
Exception for EXCEL file types: If \code{header = TRUE}, then the value \code{TRUE}
is returned.
}
\description{
Helper function for \code{read_data()}: Generate \code{col_names} object
}
