% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_col_types}
\alias{get_col_types}
\alias{get_col_types.file_structure_fwf}
\alias{get_col_types.file_structure_dsv}
\alias{get_col_types.file_structure_excel}
\title{Helper function for \code{read_data()}: Generate \code{col_types} object}
\usage{
get_col_types(obj, ...)

\method{get_col_types}{file_structure_fwf}(obj, ...)

\method{get_col_types}{file_structure_dsv}(obj, ...)

\method{get_col_types}{file_structure_excel}(obj, ...)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{...}{Optional error handler}
}
\value{
A character vector holding the type names of each column needed
for the file type specific reading function.
}
\description{
Helper function for \code{read_data()}: Generate \code{col_types} object
}
