% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_encoding}
\alias{get_encoding}
\alias{get_encoding.file_structure_fwf}
\alias{get_encoding.file_structure_dsv}
\alias{get_encoding.file_structure_sas}
\title{Helper function for \code{read_data()}: Generate encoding string}
\usage{
get_encoding(obj, ...)

\method{get_encoding}{file_structure_fwf}(obj, ...)

\method{get_encoding}{file_structure_dsv}(obj, ...)

\method{get_encoding}{file_structure_sas}(obj, ...)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{...}{Optional error handler}
}
\description{
Helper function for \code{read_data()}: Generate encoding string
}
