% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_file_type}
\alias{get_file_type}
\alias{get_file_type.file_structure_fwf}
\alias{get_file_type.file_structure_dsv}
\alias{get_file_type.file_structure_excel}
\alias{get_file_type.file_structure_sas}
\title{Get the specified file type}
\usage{
get_file_type(obj, ...)

\method{get_file_type}{file_structure_fwf}(obj, ...)

\method{get_file_type}{file_structure_dsv}(obj, ...)

\method{get_file_type}{file_structure_excel}(obj, ...)

\method{get_file_type}{file_structure_sas}(obj, ...)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{...}{Optional error handler}
}
\value{
A string holding the file type defined in \code{obj}
}
\description{
Get the specified file type
}
\section{File types}{

The function \code{\link[=read_data]{read_data()}} can read read four different types of data
\itemize{
\item \code{FWF}: Fixed width files. This files are text files, where the data is
stored in columns, that have a fixed character width.
\item \code{DSV}: Delimiter-separated value file. This files are text files, where
the data is stored in columns that are separated by a delimiter character.
\item \code{EXCEL}: An excel file holding the data.
\item \code{SAS}: A SAS file holding the data.

In order to read a data file with the function \code{\link[=read_data]{read_data()}},
it is useful to create a \link[=new_file_definition_fwf]{file_definitionuration} or
\link[=new_file_structure_fwf]{file_structure} class object,
holding all needed data file file_structures:
\itemize{
\item \code{\link[=new_file_definition_fwf]{new_file_definition_fwf()}} or \code{\link[=new_file_structure_fwf]{new_file_structure_fwf()}} for \code{FWF} files
\item \code{\link[=new_file_definition_dsv]{new_file_definition_dsv()}} or \code{\link[=new_file_structure_dsv]{new_file_structure_dsv()}} for \code{DSV} files
\item \code{\link[=new_file_definition_excel]{new_file_definition_excel()}} or \code{\link[=new_file_structure_excel]{new_file_structure_excel()}} for \code{Excel} files
\item \code{\link[=new_file_definition_sas]{new_file_definition_sas()}} or \code{\link[=new_file_structure_sas]{new_file_structure_sas()}} for \code{SAS} files
}
}
}

