% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_fwf_col_positions}
\alias{get_fwf_col_positions}
\title{Helper function for \code{read_data()}: Generate \code{col_positions} object for \code{\link[readr:read_fwf]{readr::read_fwf()}}}
\usage{
get_fwf_col_positions(
  obj,
  err_h = composerr("Error while calling 'get_fwf_col_positions()'")
)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{err_h}{An error handler}
}
\value{
An object generated by \code{readr::fwf_positions()}
}
\description{
Helper function for \code{read_data()}: Generate \code{col_positions} object for \code{\link[readr:read_fwf]{readr::read_fwf()}}
}
