% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_fwf_locale}
\alias{get_fwf_locale}
\title{Helper function for \code{read_data()}: Generate \code{locale} object for \code{\link[readr:read_fwf]{readr::read_fwf()}}}
\usage{
get_fwf_locale(obj, ...)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{...}{Optional error handler}
}
\value{
An object generated by \code{readr::default_locale()}
}
\description{
Helper function for \code{read_data()}: Generate \code{locale} object for \code{\link[readr:read_fwf]{readr::read_fwf()}}
}
