% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{get_meta_}
\alias{get_meta_}
\alias{get_meta_.file_structure}
\alias{get_meta_.data.frame}
\alias{get_meta_.file_collection}
\title{Helper function for \code{\link[=get_meta]{get_meta()}}}
\usage{
get_meta_(obj, ...)

\method{get_meta_}{file_structure}(
  obj,
  cols = NULL,
  err_h = composerr("Error while calling 'get_meta_()'"),
  ...
)

\method{get_meta_}{data.frame}(
  obj,
  cols = NULL,
  err_h = composerr("Error while calling 'get_meta_()'"),
  ...
)

\method{get_meta_}{file_collection}(
  obj,
  cols = NULL,
  err_h = composerr("Error while calling 'get_meta_()'"),
  ...
)
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class object or a
\link[=new_file_definition]{file_definition} class object or a \link[=new_file_collection]{file_collection}
class object or a \code{data.frame} generated by
the \code{\link[=read_data]{read_data()}} function (holds the meta information as
attribute). The meta information will be extracted from this
object.}

\item{...}{Additional arguments}

\item{cols}{An optional object, defining for which columns the meta
information should be retrieved. It can either be a character vector,
holding the names of the columns of interest, or it can be
numeric vector holding the indices of the columns of interest.
If \code{cols} is omitted, then all columns are used.}

\item{err_h}{An error handler}
}
\value{
Returns a \code{data.frame} holding all meta information.
}
\description{
Retrieve all meta information for a single or multiple data columns
and store it in a \code{data.frame}.
}
