% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{get_skip_rows.default}
\alias{get_skip_rows.default}
\title{Missing class for \code{\link[=get_skip_rows]{get_skip_rows()}}}
\usage{
\method{get_skip_rows}{default}(obj, err_h = composerr("Error while calling 'get_skip_rows()'"))
}
\arguments{
\item{obj}{A \link[=new_file_structure_fwf]{file_structure} class or a
\link[=new_file_definition]{file_definition} class object, that holds all informations needed for
reading a specific data set.}

\item{err_h}{An error handler}
}
\description{
Missing class for \code{\link[=get_skip_rows]{get_skip_rows()}}
}
