% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{new_adapters}
\alias{new_adapters}
\title{Create a new 'adapters' class object}
\usage{
new_adapters(..., description = NULL)
}
\arguments{
\item{...}{Either single list argument is passed to \code{new_adapters()}
(e.g. \code{new_adapters(list(fun1, fun2, ...))}), where each list argument is an
adapter function, or multiple functions are directly passed to
\code{new_adapters()} (e.g. \code{new_adapters(fun1, fun2, ...)}).}

\item{description}{An optional character vector that must have the same
length as the number of passed in adapter functions. If omitted then
non-standard-evaluation will be used in order to capture the function
name if possible. If a passed in function is not stored with a function
name, then \code{NA_character_} will be used instead.}
}
\value{
An \code{adapters} class object, holding a list of adapter functions and
their descriptions.
}
\description{
This class holds a list of adapter functions and their function description
and is usually stored inside of a \link[=new_file_definition]{file_definition} or a
\link[=new_file_structure_fwf]{file_structure} class object.
}
\section{adapters}{

An adapter function is a function that takes a data.frame as input argument
and returns a modified version of this data.frame.
The adapter functions are stored in an \link[=new_adapters]{adapters}
class object, which is a special list that contains all adapter functions
and a description text of each function. This class objects can be
created by using the function \code{\link[=new_adapters]{new_adapters()}}.
The \link[=new_adapters]{adapters} class objects can be added to a
\link[=new_file_structure_fwf]{file_structure} or a
\link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection} class object.
After reading a data file (by calling \link[=read_data]{read_data(file_definition)})
all adapter functions listed in the \code{adapters} argument of the
file_definition]\code{\link[=new_file_definition]{new_file_definition()}} class object
will be applied consecutively to the loaded data set.
Adapter functions can be added to an existing
\link[=new_file_structure_fwf]{file_structure} or a \link[=new_file_definition]{file_definition} or
a \link[=new_file_collection]{file_collection} class
object by using the function \code{\link[=add_adapters]{add_adapters()}}.
Adapter functions can be used for several tasks:
\itemize{
\item adapt the data sets in such a way that they can be concatenated for
mutliple years
\item compute new variables from existing variables
\item fix errors in variables
\item transform the values of a variable of an older data set, such that it
complies with a newer variable definition
}
}

\seealso{
\code{\link[=add_adapters]{add_adapters()}}, \code{\link[=apply_adapters]{apply_adapters()}}
}
