% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{new_col_meta}
\alias{new_col_meta}
\title{Create a new \code{col_meta} class object}
\usage{
new_col_meta(
  desc = NULL,
  values = NULL,
  values_desc = NULL,
  valid_start = NULL,
  valid_end = NULL
)
}
\arguments{
\item{desc}{An optional string holding a column description.}

\item{values}{An optional vector (character/logical/numeric) usually holding
the possible column values (e.g. \code{c(1, 2)}) or a more abstract text
version of the column values (e.g. \code{c("JJJJMMDD", "99999999", "")}).
The argument can be omitted.}

\item{values_desc}{An optional character vector that corresponds to the
\code{values} vector. Each entry of \code{values_desc} is a more detailed
description of the corresponding entry in \code{values}.
If some descriptions are not present, the entries are filled with \code{NA}.
The argument can be omitted. But if \code{values} and \code{values_desc} are both
present, then they have to be of the same length.}

\item{valid_start}{An optional atomic argument (string or number) holding
information about the first data set, in which the current
variable is valid
(present in the data sets). For example \code{valid_start = 2017}.}

\item{valid_end}{An optional atomic argument (string or number) holding
information about the last data set, in which the current
variable is valid
(present in the data sets). For example \code{valid_end = Inf}.}
}
\value{
A \code{col_meta} class object, holding the meta information for a specific
data column.
}
\description{
This class holds a set of meta information for a specific column and
is usually stored in a \link[=new_file_meta]{file_meta} class object, which is
usually stored ina \link[=new_file_structure_fwf]{file_structure} class object.
}
\section{meta information}{

The \link[=new_col_meta]{col_meta} class objects are used in order to store some
meta information about single data columns, like additional column desciptions,
and column value/level descriptions. In order to store meta information
about a set of columns a \link[=new_file_meta]{file_meta} class object can be
used. This objects store a list of \link[=new_col_meta]{col_meta} class objects, where
each \link[=new_col_meta]{col_meta} class object corresponds to a specific column in
a data set. This \link[=new_file_meta]{file_meta} class objects are usually
stored in \link[=new_file_structure_fwf]{file_structure} class objects or
\link[=new_file_definition]{file_definition} class objects. But when calling \code{\link[=read_data]{read_data()}}, the
meta information gets also appended to the resulting \code{data.frame}.
The meta information stored in a \link[=new_file_structure_fwf]{file_structure},
a \link[=new_file_definition]{file_definition} class object or a read \code{data.frame} can be extracted
by using the function \code{\link[=get_meta]{get_meta()}}.
A \link[=new_col_meta]{col_meta} class object holds the following informations:
\itemize{
\item \code{desc}: A string holding the column description.
\item \code{values}: A vector (character/logical/numeric) usually holding
the possible column values (e.g. \code{c(1, 2)}) or a more abstract text
version of the column values (e.g. \code{c("JJJJMMDD", "99999999", "")}).
\item \code{values_desc}: A character vector that corresponds to the \code{values} vector.
Each entry of \code{values_desc} is a more detailed description of the
corresponding entry in \code{values}. If some descriptions are not present,
the entries are \code{NA} values.
}
}

\seealso{
\code{\link[=new_file_meta]{new_file_meta()}}, \code{\link[=get_meta]{get_meta()}}
}
