% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_definition.R
\name{new_file_definition_}
\alias{new_file_definition_}
\title{Helper function for \code{new_file_definition()}}
\usage{
new_file_definition_(
  file_path,
  file_structure,
  to_lower,
  cols_keep,
  extra_col_name = NULL,
  extra_col_val = NULL,
  extra_col_file_path,
  extra_adapters = new_adapters(),
  err_h = composerr("Error while calling 'new_file_definition_()'")
)
}
\arguments{
\item{file_path}{A string holding the path to the data file.}

\item{file_structure}{A \link[=new_file_structure_fwf]{file_structure} class object.
This type of objects can be created by the functions
\code{\link[=new_file_structure_fwf]{new_file_structure_fwf()}}, \code{\link[=new_file_structure_dsv]{new_file_structure_dsv()}},
\code{\link[=new_file_structure_excel]{new_file_structure_excel()}} or \code{\link[=new_file_structure_sas]{new_file_structure_sas()}}
and fully defines the file structure of the
data files. The idea is that a single file_structure can be valid for
multiple data files and therefore be reused. Whereas a \code{file_definition} class
object also holds the path to the file and is therefore only valid for
a single file.}

\item{to_lower}{A logical flag, defining if the names of the columns should
be transformed to lower case after reading the data set (by calling
\code{\link[=read_data]{read_data()}}). This transformation will be applied before comparing the
column names (in the case of SAS-Files or DSV- and EXCE-Files with
\code{header = TRUE}).
In the case of \code{new_file_definition()} the \code{to_lower} argument
overwrites the \code{to_lower} argument in the
\link[=new_file_structure_fwf]{file_structure} class object given in
\code{file_structure}. If \code{to_lower} is omitted, then the \code{file_structure}
class object remains unchanged.
In the case of \code{new_file_definition_fwf()}, \code{new_file_definition_dsv()}, \code{new_file_definition_excel()}
or \code{new_file_definition_sas()} the argument \code{to_lower} must either be \code{TRUE}
or \code{FALSE}.}

\item{cols_keep}{Either \code{TRUE} or a character vector.
If set to \code{TRUE}, then all columns of the data
are kept when calling \code{\link[=read_data]{read_data()}}.
If \code{cols_keep} character vector, then the values in \code{cols_keep} represent
the names of the columns, which are kept  when calling \code{\link[=read_data]{read_data()}}.}

\item{extra_col_name}{An optional string, which defines the column, which
will be added to the data set (after reading it with function \code{\link[=read_data]{read_data()}}).
Each entry of the column will have the single value given in \code{extra_col_val}.
For example: This column is useful when reading similar data files for
separate years (one could pass the current data set year to \code{extra_col_name}
and set \code{extra_col_name = "year"}).
If \code{extra_col_name} is omitted, no column will be added to the data set and
then \code{extra_col_val} must be omitted as well.
additional column with the column name, given in  \code{extra_col_name}.
If omitted, then no column will be added to the data set and the
argument \code{extra_col_name} must be omitted as well.}

\item{extra_col_val}{An optional value (any atomic type), which will be added
(after reading the data set with function \code{\link[=read_data]{read_data()}}) as an
additional column with the column name, given in  \code{extra_col_name}.
For example: This column is useful when reading similar data files for
separate years (one could pass the current data set year to \code{extra_col_name}
and set \code{extra_col_name = "year"}).
If omitted, then no column will be added to the data set and the
argument \code{extra_col_name} must be omitted as well.}

\item{extra_col_file_path}{Either \code{FALSE} or a string.
If set to \code{FALSE} no file-path-column will be added to the data set, when
calling \code{\link[=read_data]{read_data()}}.
If the argument \code{extra_col_file_path} is a string, then a column holding
the file path of the data file will be added to the read data set, when
calling \code{\link[=read_data]{read_data()}}. The string of \code{extra_col_file_path} will be used
as column name for this additional column.}

\item{extra_adapters}{An optional \link[=new_adapters]{adapters} class object,
which holds a list of adapter functions. These adapter functions will
be added to the adapter functions already stored in the
\link[=new_file_structure_fwf]{file_structure} class object. For further details
on adapter functions see section \emph{adapters}.}

\item{err_h}{An error handler}
}
\description{
Helper function for \code{new_file_definition()}
}
