% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{new_file_meta_}
\alias{new_file_meta_}
\title{Helper function for \code{\link[=new_file_meta]{new_file_meta()}}}
\usage{
new_file_meta_(
  obj,
  err_h = composerr("Error while calling 'new_file_meta_()'")
)
}
\arguments{
\item{obj}{An unnamed list, where each entry is a \link[=new_col_meta]{col_meta} class
object.}

\item{err_h}{An error handler}
}
\description{
Create a new \link[=new_file_meta]{file_meta} class object.
}
\section{meta information}{

The \link[=new_col_meta]{col_meta} class objects are used in order to store some
meta information about single data columns, like additional column desciptions,
and column value/level descriptions. In order to store meta information
about a set of columns a \link[=new_file_meta]{file_meta} class object can be
used. This objects store a list of \link[=new_col_meta]{col_meta} class objects, where
each \link[=new_col_meta]{col_meta} class object corresponds to a specific column in
a data set. This \link[=new_file_meta]{file_meta} class objects are usually
stored in \link[=new_file_structure_fwf]{file_structure} class objects or
\link[=new_file_definition]{file_definition} class objects. But when calling \code{\link[=read_data]{read_data()}}, the
meta information gets also appended to the resulting \code{data.frame}.
The meta information stored in a \link[=new_file_structure_fwf]{file_structure},
a \link[=new_file_definition]{file_definition} class object or a read \code{data.frame} can be extracted
by using the function \code{\link[=get_meta]{get_meta()}}.
A \link[=new_col_meta]{col_meta} class object holds the following informations:
\itemize{
\item \code{desc}: A string holding the column description.
\item \code{values}: A vector (character/logical/numeric) usually holding
the possible column values (e.g. \code{c(1, 2)}) or a more abstract text
version of the column values (e.g. \code{c("JJJJMMDD", "99999999", "")}).
\item \code{values_desc}: A character vector that corresponds to the \code{values} vector.
Each entry of \code{values_desc} is a more detailed description of the
corresponding entry in \code{values}. If some descriptions are not present,
the entries are \code{NA} values.
}
}

