% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure.R
\name{new_file_structure_}
\alias{new_file_structure_}
\title{Helper function for creating a \link[=new_file_structure_fwf]{file_structure} class object}
\usage{
new_file_structure_(file_structure, class, err_h)
}
\arguments{
\item{file_structure}{A list holding all file_structures for reading an
FWF, DSV, EXCEL or SAS data file.}

\item{class}{A character vector holding the names of the subclasses of the object}

\item{err_h}{An error handler}
}
\value{
A \code{file_structure} class object
}
\description{
Helper function for creating a \link[=new_file_structure_fwf]{file_structure} class object
}
