% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_infront}
\alias{order_infront}
\title{Put some elements of a character vector infront}
\usage{
order_infront(x, vals)
}
\arguments{
\item{x}{The character vector that should be reordered}

\item{vals}{The values that should be put infront, if they are contained
in \code{x}}
}
\value{
The ordered character vector
}
\description{
Put some elements of a character vector infront
}
