% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{print.adapters}
\alias{print.adapters}
\title{Print \link[=new_adapters]{adapters} class objects}
\usage{
\method{print}{adapters}(x, indent = 0, ...)
}
\arguments{
\item{x}{The list that should be printed}

\item{indent}{A counter, defining the indentation level}

\item{...}{additional arguments}
}
\description{
Print \link[=new_adapters]{adapters} class objects
}
