% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{print.file_structure}
\alias{print.file_structure}
\title{Print \link[=new_file_structure_fwf]{file_structure} class objects}
\usage{
\method{print}{file_structure}(x, indent = 0, ...)
}
\arguments{
\item{x}{The list that should be printed}

\item{indent}{A counter, defining the indentation level}

\item{...}{additional arguments}
}
\description{
Print \link[=new_file_definition]{file_definition} class objects
}
