% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_structure_methods.R
\name{print_file_structure}
\alias{print_file_structure}
\title{Helper function for \code{\link[=print.file_structure]{print.file_structure()}} and \code{\link[=print.file_definition]{print.file_definition()}}}
\usage{
print_file_structure(x, indent = 0)
}
\arguments{
\item{x}{The list that should be printed}

\item{indent}{A counter, defining the indentation level}
}
\description{
Helper function for \code{\link[=print.file_structure]{print.file_structure()}} and \code{\link[=print.file_definition]{print.file_definition()}}
}
