% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data_.default}
\alias{read_data_.default}
\title{Missing class for \verb{[read_data_()]}}
\usage{
\method{read_data_}{default}(
  file_definition,
  err_h = composerr("Error while calling 'read_data_()'"),
  ...
)
}
\arguments{
\item{file_definition}{A file_definitionuration object, holds all informations needed for
reading the data. This object can be created with one of the following
functions:
\itemize{
\item \code{new_file_definition()}: For reading FWF, DSV or EXCEL files
\item \code{new_file_definition_fwf()}: For reading FWF files
\item \code{new_file_definition_dsv()}: For reading DSV files
\item \code{new_file_definition_excel()}: For reading EXCEL files
}}

\item{err_h}{An error handler}

\item{...}{Additional arguments}
}
\description{
Missing class for \verb{[read_data_()]}
}
