% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data_}
\alias{read_data_}
\alias{read_data_.file_collection}
\alias{read_data_.file_definition_fwf}
\alias{read_data_.file_definition_dsv}
\alias{read_data_.file_definition_excel}
\alias{read_data_.file_definition_sas}
\title{Helper method for \code{read_data()}}
\usage{
read_data_(file_definition, ...)

\method{read_data_}{file_collection}(
  file_definition,
  err_h = composerr("Error while calling 'read_data_()'"),
  try_first = TRUE
)

\method{read_data_}{file_definition_fwf}(
  file_definition,
  err_h = composerr("Error while calling 'read_data_()'")
)

\method{read_data_}{file_definition_dsv}(
  file_definition,
  err_h = composerr("Error while calling 'read_data_()'")
)

\method{read_data_}{file_definition_excel}(
  file_definition,
  err_h = composerr("Error while calling 'read_data_()'")
)

\method{read_data_}{file_definition_sas}(
  file_definition,
  err_h = composerr("Error while calling 'read_data_()'")
)
}
\arguments{
\item{file_definition}{A file_definitionuration object, holds all informations needed for
reading the data. This object can be created with one of the following
functions:
\itemize{
\item \code{new_file_definition()}: For reading FWF, DSV or EXCEL files
\item \code{new_file_definition_fwf()}: For reading FWF files
\item \code{new_file_definition_dsv()}: For reading DSV files
\item \code{new_file_definition_excel()}: For reading EXCEL files
}}

\item{...}{Additional arguments}

\item{err_h}{An error handler}

\item{try_first}{A logical value. If set to \code{TRUE}, then the function \code{\link[=read_data_]{read_data_()}}
tries reading in a small amount of rows first, in order to check
that the specified data files can be read and concatenated at all.
This should save the user from waiting a long time before getting an
error message, telling him that the file_definitionuration must be updated.}
}
\description{
Helper method for \code{read_data()}
}
