% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_null_entries}
\alias{remove_null_entries}
\title{Remove NULL entries from list}
\usage{
remove_null_entries(x)
}
\arguments{
\item{x}{A list object}
}
\value{
The list object stripped of all \code{NULL} entries
}
\description{
Remove NULL entries from list
}
