% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_adapters}
\alias{set_adapters}
\title{Set the \code{adapters} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object}
\usage{
set_adapters(obj, value)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{adapters} argument should be set.}

\item{value}{An \link[=new_adapters]{adapters} class object}
}
\value{
The modified object
}
\description{
Set the \code{adapters} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object
}
