% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_adapters_.default}
\alias{set_adapters_.default}
\title{Wrong class case for \code{set_adapters_()}}
\usage{
\method{set_adapters_}{default}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{adapters} argument should be set.}

\item{value}{An \link[=new_adapters]{adapters} class object}

\item{err_h}{An error handler}
}
\description{
Wrong class case for \code{set_adapters_()}
}
