% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_adapters_}
\alias{set_adapters_}
\alias{set_adapters_.file_structure}
\alias{set_adapters_.file_collection}
\title{Helper function for \code{\link[=set_adapters]{set_adapters()}}}
\usage{
set_adapters_(obj, value, err_h)

\method{set_adapters_}{file_structure}(obj, value, err_h)

\method{set_adapters_}{file_collection}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{adapters} argument should be set.}

\item{value}{An \link[=new_adapters]{adapters} class object}

\item{err_h}{An error handler}
}
\description{
Helper function for \code{\link[=set_adapters]{set_adapters()}}
}
