% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_cols_keep}
\alias{set_cols_keep}
\title{Set the \code{cols_keep} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object}
\usage{
set_cols_keep(obj, value)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{cols_keep} argument should be set.}

\item{value}{A character vector holding the column names, which should
be kept.}
}
\value{
The modified object
}
\description{
Set the \code{cols_keep} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object
}
