% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_cols_keep_.default}
\alias{set_cols_keep_.default}
\title{Wrong class case for \code{set_cols_keep_()}}
\usage{
\method{set_cols_keep_}{default}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{cols_keep} argument should be set.}

\item{value}{A character vector holding the column names, which should
be kept.}

\item{err_h}{An error handler}
}
\description{
Wrong class case for \code{set_cols_keep_()}
}
