% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_cols_keep_}
\alias{set_cols_keep_}
\alias{set_cols_keep_.file_definition}
\alias{set_cols_keep_.file_collection}
\title{Helper function for \code{\link[=set_cols_keep]{set_cols_keep()}}}
\usage{
set_cols_keep_(obj, value, err_h)

\method{set_cols_keep_}{file_definition}(obj, value, err_h)

\method{set_cols_keep_}{file_collection}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{cols_keep} argument should be set.}

\item{value}{A character vector holding the column names, which should
be kept.}

\item{err_h}{An error handler}
}
\description{
Helper function for \code{\link[=set_cols_keep]{set_cols_keep()}}
}
