% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_extra_col_file_path}
\alias{set_extra_col_file_path}
\title{Set the \code{extra_col_file_path} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object}
\usage{
set_extra_col_file_path(obj, value)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{extra_col_file_path} argument should be set.}

\item{value}{A string holding the column name, of the column holding the
file path strings.}
}
\value{
The modified object
}
\description{
Set the \code{extra_col_file_path} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object
}
