% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_extra_col_file_path_.default}
\alias{set_extra_col_file_path_.default}
\title{Wrong class case for \code{set_extra_col_file_path_()}}
\usage{
\method{set_extra_col_file_path_}{default}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{extra_col_file_path} argument should be set.}

\item{value}{A string holding the column name, of the column holding the
file path strings.}

\item{err_h}{An error handler}
}
\description{
Wrong class case for \code{set_extra_col_file_path_()}
}
