% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_extra_col_file_path_}
\alias{set_extra_col_file_path_}
\alias{set_extra_col_file_path_.file_definition}
\alias{set_extra_col_file_path_.file_collection}
\title{Helper function for \code{\link[=set_extra_col_file_path]{set_extra_col_file_path()}}}
\usage{
set_extra_col_file_path_(obj, value, err_h)

\method{set_extra_col_file_path_}{file_definition}(obj, value, err_h)

\method{set_extra_col_file_path_}{file_collection}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{extra_col_file_path} argument should be set.}

\item{value}{A string holding the column name, of the column holding the
file path strings.}

\item{err_h}{An error handler}
}
\description{
Helper function for \code{\link[=set_extra_col_file_path]{set_extra_col_file_path()}}
}
