% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_n_max}
\alias{set_n_max}
\title{Set the \code{n_max} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object}
\usage{
set_n_max(obj, value)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{n_max} argument should be set.}

\item{value}{Either an integer defining the maximum number of rows, which
should be read when calling \code{\link[=read_data]{read_data()}}.}
}
\value{
The modified object
}
\description{
Set the \code{n_max} argument for a \link[=new_file_definition]{file_definition} or a
\link[=new_file_collection]{file_collection} class object
}
