% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_n_max_.default}
\alias{set_n_max_.default}
\title{Wrong class case for \code{set_n_max_()}}
\usage{
\method{set_n_max_}{default}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{n_max} argument should be set.}

\item{value}{Either an integer defining the maximum number of rows, which
should be read when calling \code{\link[=read_data]{read_data()}}.}

\item{err_h}{An error handler}
}
\description{
Wrong class case for \code{set_n_max_()}
}
