% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{set_n_max_}
\alias{set_n_max_}
\alias{set_n_max_.file_structure}
\alias{set_n_max_.file_collection}
\title{Helper function for \code{\link[=set_n_max]{set_n_max()}}}
\usage{
set_n_max_(obj, value, err_h)

\method{set_n_max_}{file_structure}(obj, value, err_h)

\method{set_n_max_}{file_collection}(obj, value, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} or a \link[=new_file_collection]{file_collection}
class object, for which the \code{n_max} argument should be set.}

\item{value}{Either an integer defining the maximum number of rows, which
should be read when calling \code{\link[=read_data]{read_data()}}.}

\item{err_h}{An error handler}
}
\description{
Helper function for \code{\link[=set_n_max]{set_n_max()}}
}
