% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stringify}
\alias{stringify}
\title{convert vector to string}
\usage{
stringify(x)
}
\arguments{
\item{x}{A numeric or character vector}
}
\value{
A string
}
\description{
e.g. `c(1, 2, 3) -> "'1', '2', '3'"
}
