% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{type_of_var}
\alias{type_of_var}
\title{Get type of object, but also distinguishing \code{factor}}
\usage{
type_of_var(x)
}
\arguments{
\item{x}{An object}
}
\value{
Possible strings:
\itemize{
\item \code{"logical"}
\item \code{"numeric"}
\item \code{"character"}
\item \code{"factor"}
\item \code{"list"}
\item various other non-atomic types
}
}
\description{
Get type of object, but also distinguishing \code{factor}
}
