% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_collection.R
\name{validate_file_collection}
\alias{validate_file_collection}
\title{Validate \link[=new_file_collection]{file_collection} object}
\usage{
validate_file_collection(file_collection, validate_class = TRUE, err_h)
}
\arguments{
\item{file_collection}{A \link[=new_file_collection]{file_collection} object}

\item{validate_class}{A logical argument, defining if the object class \code{file_collection}
required.}

\item{err_h}{An error handler}
}
\description{
Validate \link[=new_file_collection]{file_collection} object
}
