% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file_definition.R
\name{validate_file_definition}
\alias{validate_file_definition}
\title{Validate \link[=new_file_definition]{file_definition} object}
\usage{
validate_file_definition(obj, err_h)
}
\arguments{
\item{obj}{A \link[=new_file_definition]{file_definition} object}

\item{err_h}{An error handler}
}
\description{
Validate \link[=new_file_definition]{file_definition} object
}
