% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file_structure.R
\name{validate_file_structure.default}
\alias{validate_file_structure.default}
\title{Missing class for \code{\link[=validate_file_structure]{validate_file_structure()}}}
\usage{
\method{validate_file_structure}{default}(obj, err_h)
}
\arguments{
\item{obj}{A \code{file_definition} object}

\item{err_h}{An error handler}
}
\description{
Missing class for \code{\link[=validate_file_structure]{validate_file_structure()}}
}
