% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file_structure.R
\name{validate_file_structure}
\alias{validate_file_structure}
\alias{validate_file_structure.file_structure}
\alias{validate_file_structure.file_structure_fwf}
\alias{validate_file_structure.file_structure_dsv}
\alias{validate_file_structure.file_structure_excel}
\alias{validate_file_structure.file_structure_sas}
\title{Validate \code{file_structure} object}
\usage{
validate_file_structure(obj, ...)

\method{validate_file_structure}{file_structure}(
  obj,
  err_h = composerr("Error while calling 'validate_file_structure()'")
)

\method{validate_file_structure}{file_structure_fwf}(
  obj,
  err_h = composerr("Error while calling 'validate_file_structure()'")
)

\method{validate_file_structure}{file_structure_dsv}(
  obj,
  err_h = composerr("Error while 'validate_file_structure()'")
)

\method{validate_file_structure}{file_structure_excel}(
  obj,
  err_h = composerr("Error while calling 'validate_file_structure()'")
)

\method{validate_file_structure}{file_structure_sas}(
  obj,
  err_h = composerr("Error while calling 'validate_file_structure()'")
)
}
\arguments{
\item{obj}{A \code{file_definition} object}

\item{...}{Additional arguments}

\item{err_h}{An error handler}
}
\description{
Validate \code{file_structure} object
}
